+++ src/app/features/forms/swap-form/utils/getExistingPositionFixedRate/index.ts
@@ -1,6 +1,6 @@
 import { Draft } from '@reduxjs/toolkit';
 
-import { SliceState as CashflowCalculatorState } from '../../../../cashflow-calculator/reducer';
+import { SliceState as CashflowCalculatorState } from '../../../../cashflow-calculator/state';
 import { SliceState as SwapFormState } from '../../state';
 import { getEditPositionNotional } from '../getEditPositionNotional';
